Enable calling the indexers as functions returns promises
Use require instead of include

Use the native promise instead of bluebird

Added server for node apps will add the required routes

Fixing server
