+++ lib/cli/options-formatter.js
@@ -1,7 +1,7 @@
 /**
  * Dependencies.
  */
-var utils = include('utils');
+var utils = require('../utils');
 
 /**
  * Option parser: this module provides
