+++ lib/command/bootstrap.js
@@ -1,10 +1,10 @@
-var logger = include('logger');
-var utils = include('utils');
-var config = include('config');
+var logger = require('../logger');
+var utils = require('../utils');
+var config = require('../config');
 var entitiesPath = config.get('indexer.entities.directory');
 var _ = require('lodash');
 var path = require('path');
-var optionsFormatter = include('cli/options-formatter');
+var optionsFormatter = require('../cli/options-formatter');
 
 /**
  * Index command.
