+++ lib/command/index-realtime.js
@@ -1,9 +1,9 @@
 /**
  * Dependencies.
  */
-var logger = include('logger');
-var indexer = include('command/index');
-var amqp = include('storages/amqp');
+var logger = require('../logger');
+var indexer = require('./index');
+var amqp = require('../storages/amqp');
 var _ = require('lodash');
 
 /**
