+++ lib/command/watch.js
@@ -1,12 +1,12 @@
-var optionsFormatter = include('cli/options-formatter');
-var config = include('config');
+var optionsFormatter = require('../cli/options-formatter');
+var config = require('../config');
 var defaultInterval = config.get('watcher.defaultInterval');
-var logger = include('logger');
+var logger = require('../logger');
 var fs = require('fs');
 var path = require('path');
 var watchersDirectory = config.get('dirs.watchers') || path.join(config.get('dirs.base') || config.get('rootDir'), 'watchers');
 var _ = require('lodash');
-var notifier = include('notifier');
+var notifier = require('../notifier');
 var watchers = [];
 
 if (!fs.existsSync(watchersDirectory)) {
