+++ lib/notifier.js
@@ -1,7 +1,6 @@
-var amqp = include('storages/amqp');
-var logger = include('logger');
-var utils = include('utils');
-var Promise = require('bluebird');
+var amqp = require('./storages/amqp');
+var logger = require('./logger');
+var utils = require('./utils');
 
 var Notifier = function(options) {
   this.options = options;
