+++ lib/storages/amqp.js
@@ -1,7 +1,6 @@
 var amqp = require('amqp');
-var Promise = require('bluebird');
-var config = include('config');
-var logger = include('logger');
+var config = require('../config');
+var logger = require('../logger');
 var amqpServerConfig = config.get('amqp');
 var connectionRetry = config.get('amqp.retry');
