+++ lib/storages/mysql.js
@@ -1,8 +1,7 @@
-var Promise = require('bluebird');
-var config = include('config');
+var config = require('../config');
 var mysqlLib = require('mysql2');
 var serversConfig = config.get('mysql');
-var logger = include('logger');
+var logger = require('../logger');
 var _ = require('lodash');
 var dbPools = {};
 var vpo = require('vpo');
