+++ lib/utils.js
@@ -1,7 +1,6 @@
 /**
  * Dependencies, it includes the node's util lib.
  */
-var Promise = require('bluebird');
 var path = require('path');
 var fs = require('fs');
 var _ = require('lodash');
@@ -38,7 +37,7 @@ utils.getResources = function(directory, filter) {
   return new Promise(function(resolve, reject) {
     fs.readdir(directory, function(err, resources) {
       if (err) {
-        var logger = include('logger');
+        var logger = require('./logger');
         logger.error('You are requesting files inside ' + directory + ' but the directory doesn\'t exist');
         logger.error('Perhaps you are trying to setup new sources / backends? Then create that folder ;-)');
